requirements.txt
django==3.0.8
django-extensions==3.0.3
django-debug-toolbar==2.2